-changes in "Testing and analyzing.ipynb": trying to distort image, first volume estimation -changes in "exif_extract": checks first if depthmap already exists

+++ exif_extract.py
@@ -1,3 +1,5 @@
+import pathlib
+
 from exiftool import ExifToolHelper, ExifTool
 import os
 from urllib.parse import urlparse
@@ -42,28 +44,33 @@ def transform_jpeg_to_depthmap_samsungs20(url: str, output_dir: str = None) -> t
 
     picture_name = input_path.stem
     output_dir = Path(output_dir or "SamsungTiefenkarten")
-    output_dir.mkdir(parents=True, exist_ok=True)  # Zielordner erstellen, falls nicht vorhanden
-    output_path = output_dir / f"{picture_name}_DepthMap.tiff"
+    #output_dir.mkdir(parents=True, exist_ok=True)  # Zielordner erstellen, falls nicht vorhanden
+    output_path = output_dir / f"{picture_name}_DepthMap.jpg"
 
-    # ExifTool verwenden, um die Tiefenkarte zu extrahieren
-    try:
-        with ExifTool() as et:
-            et.execute("-b", "-DepthMapTiff", str(input_path), raw_bytes=True)
-            binary_data = et.last_stdout
-    except Exception as e:
-        raise RuntimeError(f"Fehler beim Ausführen von ExifTool: {e}")
+    file_exists = pathlib.Path.exists(output_path)
 
-    # Überprüfen, ob Daten extrahiert wurden
-    if binary_data:
+    if file_exists == False:
+        # ExifTool verwenden, um die Tiefenkarte zu extrahieren
         try:
-            with open(output_path, "wb") as f:
-                f.write(binary_data)
-            print(f"Tiefenkarte erfolgreich gespeichert: {output_path}")
+            with ExifTool() as et:
+                et.execute("-b", "-DepthMapTiff", str(input_path), raw_bytes=True)
+                binary_data = et.last_stdout
         except Exception as e:
-            raise RuntimeError(f"Fehler beim Schreiben der Tiefenkarte: {e}")
+            raise RuntimeError(f"Fehler beim Ausführen von ExifTool: {e}")
+
+        # Überprüfen, ob Daten extrahiert wurden
+        if binary_data:
+            try:
+                with open(output_path, "wb") as f:
+                    f.write(binary_data)
+                print(f"Tiefenkarte erfolgreich gespeichert: {output_path}")
+            except Exception as e:
+                raise RuntimeError(f"Fehler beim Schreiben der Tiefenkarte: {e}")
+        else:
+            print("Keine Tiefenkarte gefunden oder extrahiert.")
+            return url, None
     else:
-        print("Keine Tiefenkarte gefunden oder extrahiert.")
-        return url, None
+        print("Die Tiefenkarte existiert bereits")
 
     return str(input_path), str(output_path)
 
@@ -76,39 +83,36 @@ def transform_jpeg_to_depthmap_iphone(url: str, output_dir: str = None) -> tuple
 
     picture_name = input_path.stem
     output_dir = Path(output_dir or "IPhoneTiefenkarten")
-    output_dir.mkdir(parents=True, exist_ok=True)  # Zielordner erstellen, falls nicht vorhanden
+    #output_dir.mkdir(parents=True, exist_ok=True)  # Zielordner erstellen, falls nicht vorhanden
     output_path = output_dir / f"{picture_name}_DepthMap.jpg"
 
-    # ExifTool verwenden, um die Tiefenkarte zu extrahieren
-    try:
-        with ExifTool() as et:
-            et.execute("-b", "-MPImage3", str(input_path), raw_bytes=True)
-            binary_data = et.last_stdout
-    except Exception as e:
-        raise RuntimeError(f"Fehler beim Ausführen von ExifTool: {e}")
+    file_exists = pathlib.Path.exists(output_path)
+    print(file_exists)
 
-    # Überprüfen, ob Daten extrahiert wurden
-    if binary_data:
+    if file_exists == False:
+        # ExifTool verwenden, um die Tiefenkarte zu extrahieren
         try:
-            with open(output_path, "wb") as f:
-                f.write(binary_data)
-            print(f"Tiefenkarte erfolgreich gespeichert: {output_path}")
+            with ExifTool() as et:
+                et.execute("-b", "-MPImage3", str(input_path), raw_bytes=True)
+                binary_data = et.last_stdout
         except Exception as e:
-            raise RuntimeError(f"Fehler beim Schreiben der Tiefenkarte: {e}")
+            raise RuntimeError(f"Fehler beim Ausführen von ExifTool: {e}")
+
+        # Überprüfen, ob Daten extrahiert wurden
+        if binary_data:
+            try:
+                with open(output_path, "wb") as f:
+                    f.write(binary_data)
+                print(f"Tiefenkarte erfolgreich gespeichert: {output_path}")
+            except Exception as e:
+                raise RuntimeError(f"Fehler beim Schreiben der Tiefenkarte: {e}")
+        else:
+            print("Keine Tiefenkarte gefunden oder extrahiert.")
+            return url, None
     else:
-        print("Keine Tiefenkarte gefunden oder extrahiert.")
-        return url, None
+        print("Die Tiefenkarte existiert bereits")
 
     return str(input_path), str(output_path)
 
 
 # ------Aufruf von Funktionen------
-url = r"C:\Users\Diren\Nextcloud\HTW\4.Semester-Masterarbeit\Masterarbeit\Code\SamsungTestBilder\20250105_164600.jpg"
-output_dir = r"C:\Users\Diren\Nextcloud\HTW\4.Semester-Masterarbeit\Masterarbeit\Code\Tiefenkarten"
-
-url_iphone = r"C:\Users\Diren\Nextcloud\HTW\4.Semester-Masterarbeit\Masterarbeit\Code\IphoneTestBilder\IMG_8173.jpg"
-output_dir_iphone = r"C:\Users\Diren\Nextcloud\HTW\4.Semester-Masterarbeit\Masterarbeit\Code\IPhoneTiefenkarten"
-
-input_path, output_path = transform_jpeg_to_depthmap_samsungs20(url, output_dir)
-
-input_path_iphone, output_path_iphone = transform_jpeg_to_depthmap_iphone(url_iphone, output_dir_iphone)
